+++ examples/counter/src/main.rs
@@ -5,7 +5,6 @@ use bevy::window::WindowTheme;
 use bevy_cobweb::prelude::*;
 use bevy_cobweb_ui::prelude::*;
 use bevy_cobweb_ui::sickle::ui_builder::*;
-use bevy_cobweb_ui::sickle::SickleUiPlugin;
 
 //-------------------------------------------------------------------------------------------------------------------
 
@@ -66,8 +65,6 @@ fn main()
             primary_window: Some(Window { window_theme: Some(WindowTheme::Dark), ..default() }),
             ..default()
         }))
-        .add_plugins(SickleUiPlugin)
-        .add_plugins(ReactPlugin)
         .add_plugins(CobwebUiPlugin)
         .load("main.caf.json")
         .add_systems(PreStartup, setup)